chore: replace old experiment detail page with hp viz experiment detail page
